some fix...

+++ server/src/cron/services/cron.service.ts
@@ -1,10 +1,13 @@
 import { Injectable } from '@nestjs/common';
 import { Cron } from '@nestjs/schedule';
 import * as fs from 'fs';
+import { ImageService } from '../../images/services';
 
 @Injectable()
 export class CronService {
-  @Cron('0 3 * * * *')
+  constructor(private readonly imageService: ImageService) {}
+
+  @Cron('0 0 7 * * *')
   public cleanupImages() {
     this.cleanup();
   }
@@ -12,8 +15,9 @@ export class CronService {
   private cleanup() {
     const uploadDir = 'uploads/';
     const files = fs.readdirSync(uploadDir, { recursive: true });
+    const isUploaded = this.imageService.isUploaded;
 
-    if (fs.existsSync(uploadDir) && files.length) {
+    if (fs.existsSync(uploadDir) && files.length && isUploaded) {
       for (const file of files) {
         fs.unlinkSync(uploadDir + file);
       }
